Started on arg asserts
