+++ tests/conflicts.rs
@@ -127,3 +127,80 @@ fn conflict_with_unused_default_value() {
     assert_eq!(m.value_of("opt"), Some("default"));
     assert!(m.is_present("flag"));
 }
+
+#[test]
+fn two_conflicting_arguments() {
+    let a = App::new("two_conflicting_arguments")
+        .arg(
+            Arg::with_name("develop")
+                .long("develop")
+                .conflicts_with("production"),
+        )
+        .arg(
+            Arg::with_name("production")
+                .long("production")
+                .conflicts_with("develop"),
+        )
+        .try_get_matches_from(vec!["", "--develop", "--production"]);
+
+    assert!(a.is_err());
+    let a = a.unwrap_err();
+    assert_eq!(
+        a.cause,
+        "The argument \'--develop\' cannot be used with \'--production\'"
+    );
+}
+
+#[test]
+fn three_conflicting_arguments() {
+    let a = App::new("two_conflicting_arguments")
+        .arg(
+            Arg::with_name("one")
+                .long("one")
+                .conflicts_with_all(&["two", "three"]),
+        )
+        .arg(
+            Arg::with_name("two")
+                .long("two")
+                .conflicts_with_all(&["one", "three"]),
+        )
+        .arg(
+            Arg::with_name("three")
+                .long("three")
+                .conflicts_with_all(&["one", "two"]),
+        )
+        .try_get_matches_from(vec!["", "--one", "--two", "--three"]);
+
+    assert!(a.is_err());
+    let a = a.unwrap_err();
+    assert_eq!(
+        a.cause,
+        "The argument \'--one\' cannot be used with \'--two\'"
+    );
+}
+
+#[cfg(debug_assertions)]
+#[test]
+#[should_panic(expected = "Argument 'config' cannot conflict with itself")]
+fn self_conflicting_arg() {
+    let _ = App::new("prog")
+        .arg(
+            Arg::with_name("config")
+                .long("config")
+                .conflicts_with("config"),
+        )
+        .try_get_matches_from(vec!["", "--config"]);
+}
+
+#[cfg(debug_assertions)]
+#[test]
+#[should_panic(expected = "Argument specified in 'conflicts_with*' for 'config' does not exist")]
+fn conflicts_with_invalid_arg() {
+    let _ = App::new("prog")
+        .arg(
+            Arg::with_name("config")
+                .long("config")
+                .conflicts_with("extra"),
+        )
+        .try_get_matches_from(vec!["", "--config"]);
+}
